Moved DistributeUtil to Model Component. Moved DistributeUtilTest to ModelTest Component. All files commited are affected by the refactoring. Redesigned Sequence Diagram and UML Class diagram due to restructuring
