+++ src/main/java/seedu/address/model/distribute/DistributeAlgorithm.java
@@ -15,7 +15,6 @@ import java.util.logging.Logger;
 
 import javafx.collections.ObservableList;
 import seedu.address.commons.core.LogsCenter;
-import seedu.address.commons.util.DistributeUtil;
 import seedu.address.logic.commands.exceptions.CommandException;
 import seedu.address.model.Model;
 import seedu.address.model.person.Nationality;
