+++ src/main/java/seedu/address/model/distribute/DistributeUtil.java
@@ -1,4 +1,4 @@
-package seedu.address.commons.util;
+package seedu.address.model.distribute;
 
 import static java.util.Objects.requireNonNull;
 import static java.util.stream.Collectors.toMap;
